\name{loadtest.wipe}
\alias{loadtest.wipe}
\title{Delete campaign generated by loadtest function}
\usage{
  loadtest.wipe(n.users = 10,
    xmlfile = system.file(package = "Ohmage", "files/jeroen.xml"),
    recycle = TRUE)
}
\arguments{
  \item{n.users}{number of users that was generated}

  \item{xmlfile}{path or string of orriginal xml}

  \item{recycle}{to recycle the connection}
}
\description{
  Delete campaign generated by loadtest function
}

