\name{oh.campaign.read}
\alias{oh.campaign.read}
\title{List campaigns available on the server}
\usage{
  oh.campaign.read(output_format = "short",
    to.data.frame = TRUE, ...)
}
\arguments{
  \item{output_format}{short or long}

  \item{to.data.frame}{should response be converted to a
  dataframe}

  \item{...}{other arguments passed to ohmage}
}
\description{
  List campaigns available on the server
}

