\name{oh.class.create}
\alias{oh.class.create}
\title{Create a new class}
\usage{
  oh.class.create(class_urn, class_name,
    description = "My new class.", ...)
}
\arguments{
  \item{class_urn}{class id}

  \item{class_name}{class name}

  \item{description}{class description}

  \item{...}{other arguments passed to ohmage}
}
\description{
  Create a new class
}

