\name{oh.getdata}
\alias{oh.getdata}
\title{Depricated wrapper for oh.survey_response.read}
\usage{
  oh.getdata(serverurl, token, ...)
}
\arguments{
  \item{serverurl}{url of server}

  \item{token}{token}

  \item{...}{other args}
}
\description{
  Depricated wrapper for oh.survey_response.read
}

