\name{oh.image.read}
\alias{oh.image.read}
\title{Download an image}
\usage{
  oh.image.read(campaign_urn, owner, id, ...)
}
\arguments{
  \item{campaign_urn}{campaign id}

  \item{owner}{owner of the image}

  \item{id}{photo uuid}

  \item{...}{other parameters}
}
\value{
  path to image
}
\description{
  Download an image
}

