\name{loadtest.wipe}
\alias{loadtest.wipe}
\title{Delete campaign generated by loadtest function}
\usage{
  loadtest.wipe(n.users = 10,
    xmlfile = system.file(package = "Ohmage", "files/jeroen.xml"),
    recycle = TRUE, user.prefix = "loadtest")
}
\arguments{
  \item{n.users}{number of users that was generated}

  \item{xmlfile}{path or string of orriginal xml}

  \item{recycle}{to recycle the connection}

  \item{user.prefix}{a character string used in the
  usernames and campaign name. Make it short.}
}
\description{
  Delete campaign generated by loadtest function
}

