\name{loadtest.wipe}
\alias{loadtest.wipe}
\title{Delete campaign generated by loadtest function}
\usage{
  loadtest.wipe(xmlfile, n.users, recycle = TRUE)
}
\arguments{
  \item{xmlfile}{path or string of orriginal xml}

  \item{n.users}{number of users that was generated}

  \item{recycle}{to recycle the connection}
}
\description{
  Delete campaign generated by loadtest function
}

