\name{oh.login}
\alias{oh.login}
\title{Authenticate with an ohmage server}
\usage{
  oh.login(user, password, serverurl, ...)
}
\arguments{
  \item{user}{ohmage username}

  \item{password}{ohmage passwd}

  \item{serverurl}{url to the ohmage server}

  \item{...}{extra parameters for oh.call}
}
\value{
  null
}
\description{
  Authenticate with an ohmage server
}
\author{
  jeroen
}

